requirements.txt
# Friccion - dependencias completas (pasos 1-5 + web + heartbeats + busqueda)
openai>=1.0.0
python-dotenv>=1.0.0
python-telegram-bot>=21.0
flask>=3.0.0
flask-cors>=4.0.0
schedule>=1.2.0
duckduckgo-search>=7.0.0
tiktoken>=0.5.0
